hitcon-quals/2025/imgc0nv/requirements.txt
Flask==3.1.2
gunicorn==23.0.0
pillow==11.3.0